requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.1
django-filter==22.1
djangorestframework==3.13.1
importlib-metadata==4.12.0
Markdown==3.4.1
psycopg2==2.9.3
pytz==2022.2.1
sqlparse==0.4.2
zipp==3.8.1
